@@ -1,20 +1,19 @@
-package com.akabex86.commands;
+package com.akabex86.features.warps.commands;
 
-import com.akabex86.main.Main;
-import com.akabex86.utils.Warp;
+import java.util.ArrayList;
+import java.util.List;
+
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
 import org.bukkit.entity.Player;
 
-import java.util.ArrayList;
-import java.util.List;
+import com.akabex86.features.warps.WarpManager;
 
 public class CommandWarp implements CommandExecutor, TabCompleter {
-    public CommandWarp(Main main){
-
-    }
+	
+    
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String arg2, String[] args) {
 
@@ -24,11 +23,11 @@ public class CommandWarp implements CommandExecutor, TabCompleter {
                 p.sendMessage("Fehler: Bitte gib einen Namen für deinen Warp an.");
                 return true;
             }
-            if(Warp.getWarp(args[0].toLowerCase())==null){
+            if(WarpManager.getWarp(args[0].toLowerCase())==null){
                 p.sendMessage("Fehler: Warp ["+args[0]+"] nicht gefunden.");
                 return true;
             }
-            p.teleport(Warp.getWarp(args[0].toLowerCase()));
+            p.teleport(WarpManager.getWarp(args[0].toLowerCase()));
             p.sendMessage("Teleportiere zum Warp ["+args[0]+"] ...");
             return true;
         }
@@ -38,11 +37,9 @@ public class CommandWarp implements CommandExecutor, TabCompleter {
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String s, String[] strings) {
         List<String> TabComplete = new ArrayList<>();
-        if (sender instanceof Player){
-
-            Player p = (Player)sender;
-
-            TabComplete.addAll(Warp.getWarps().keySet());
+        if (sender instanceof Player p){
+            //TODO: Add logic to this
+            TabComplete.addAll(WarpManager.getWarps().keySet());
         }
         return TabComplete;
     }
