@@ -1,10 +1,8 @@
-package com.akabex86.commands;
+package com.akabex86.features.warps.commands;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
-import org.bukkit.Location;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -13,23 +11,18 @@ import org.bukkit.entity.Player;
 
 import com.akabex86.features.warps.WarpManager;
 
-public class CommandSpawn implements CommandExecutor, TabCompleter {
-	
-	private static final String SPAWNKEY = "spawn";
+public class CommandWarps implements CommandExecutor, TabCompleter {
 	
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String arg2, String[] args) {
     	if(sender instanceof Player player) {
-    		Optional<Location> spawnLocation = Optional.ofNullable(WarpManager.getWarp(SPAWNKEY));
-    		
-    		spawnLocation.ifPresentOrElse(location -> {
-    			player.teleport(location);
-    			player.sendMessage("Teleportiere zum Spawn ...");
-    		}, () ->  player.sendMessage("Fehler: Spawnpunkt nicht gefunden."));
+    		StringBuilder textMessage = new StringBuilder("Warps: ");
+    		WarpManager.getWarps().keySet().forEach(warp -> textMessage.append(String.format("%n- %s", warp)));
+    		player.sendMessage(textMessage.toString());
+    		return true;
     	}
-        return true;
+    	return false;
     }
-
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String s, String[] strings) {
         return Collections.emptyList();
