@@ -4,12 +4,18 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.bukkit.Bukkit;
+
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
 import org.bukkit.plugin.java.JavaPlugin;
 
 import com.akabex86.commands.CommandBook;
 import com.akabex86.commands.CommandBroadcast;
 import com.akabex86.commands.CommandDelhome;
-import com.akabex86.commands.CommandDelwarp;
+
 import com.akabex86.commands.CommandEnchant;
 import com.akabex86.commands.CommandEnderchest;
 import com.akabex86.commands.CommandHome;
@@ -17,22 +23,23 @@ import com.akabex86.commands.CommandHomes;
 import com.akabex86.commands.CommandInvsee;
 import com.akabex86.commands.CommandPlayerinfo;
 import com.akabex86.commands.CommandSethome;
-import com.akabex86.commands.CommandSetwarp;
+
 import com.akabex86.commands.CommandSpawn;
 import com.akabex86.commands.CommandSpyhome;
 import com.akabex86.commands.CommandTicket;
 import com.akabex86.commands.CommandTpa;
 import com.akabex86.commands.CommandTpaccept;
 import com.akabex86.commands.CommandTpdeny;
-import com.akabex86.commands.CommandWarp;
-import com.akabex86.commands.CommandWarps;
+
 import com.akabex86.commands.CommandZone;
 import com.akabex86.features.FeatureManager;
+import com.akabex86.features.dailyrewards.LoginTracker;
+
 import com.akabex86.listeners.PacketListeners;
 import com.akabex86.listeners._EventLoader;
 import com.akabex86.utils.Config;
 import com.akabex86.utils.UuidFetcher;
-import com.akabex86.utils.Warp;
+
 import com.comphenix.protocol.ProtocolLibrary;
 import com.comphenix.protocol.ProtocolManager;
 
@@ -58,6 +65,8 @@ public class Main extends JavaPlugin {
             getLogger().warning("PlaceholderAPI nicht gefunden! Plugin texte werden eventuell fehlerhaft dargestellt!");
         }
 
+      
+        
         protocolManager = ProtocolLibrary.getProtocolManager();
 
         loadConfig();
@@ -66,7 +75,6 @@ public class Main extends JavaPlugin {
         loadListeners();
         PacketListeners.loadPacketListeners(protocolManager);
 
-        Warp.loadWarps();//Only works properly if the file exists, FIX THAT!
         UuidFetcher.updateMappings();
         
         /*
@@ -78,6 +86,17 @@ public class Main extends JavaPlugin {
         featureManager.initializeFeatures();
         logger.log(Level.INFO, "All Features loaded!");
         
+
+        getCommand("checkLastLogin").setExecutor(new CommandExecutor() {
+			
+			@Override
+			public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+				sender.sendMessage(LoginTracker.getLastLoginMilis((Player) sender)+"");
+				return true;
+			}
+		});
+        
+
         logger.log(Level.INFO,"Survival System aktiviert!");
     }
     @Override
@@ -111,25 +130,9 @@ public class Main extends JavaPlugin {
         getCommand("delhome").setTabCompleter(delhome);
 
         //WARP COMMANDS
-        CommandSpawn spawn = new CommandSpawn(this);
+        CommandSpawn spawn = new CommandSpawn();
         getCommand("spawn").setExecutor(spawn);
         getCommand("spawn").setTabCompleter(spawn);
-
-        CommandWarp warp = new CommandWarp(this);
-        getCommand("warp").setExecutor(warp);
-        getCommand("warp").setTabCompleter(warp);
-
-        CommandWarps warps = new CommandWarps(this);
-        getCommand("warps").setExecutor(warps);
-        getCommand("warps").setTabCompleter(warps);
-
-        CommandSetwarp setwarp = new CommandSetwarp(this);
-        getCommand("setwarp").setExecutor(setwarp);
-        getCommand("setwarp").setTabCompleter(setwarp);
-
-        CommandDelwarp delwarp = new CommandDelwarp(this);
-        getCommand("delwarp").setExecutor(delwarp);
-        getCommand("delwarp").setTabCompleter(delwarp);
 
         //TPA COMMANDS
         CommandTpa tpa = new CommandTpa(this);
@@ -173,8 +176,10 @@ public class Main extends JavaPlugin {
         CommandBroadcast bc= new CommandBroadcast(this);
         getCommand("broadcast").setExecutor(bc);
         getCommand("broadcast").setTabCompleter(bc);
+       
 
     }
+    
     private void loadConfig(){
         Config.createMainFolder();
         Config.createFolder("kits");
